Clean up project directory structure
