+++ Package.swift
@@ -10,7 +10,8 @@ let package = Package(
     products: [
         .library(
             name: "TurboNavigator",
-            targets: ["TurboNavigator"]),
+            targets: ["TurboNavigator"]
+        ),
     ],
     dependencies: [
         .package(url: "https://github.com/hotwired/turbo-ios", revision: "2ace9ebe01fe3bce3e7e8df38cb82df052493663"),
@@ -20,8 +21,13 @@ let package = Package(
             name: "TurboNavigator",
             dependencies: [
                 .product(name: "Turbo", package: "turbo-ios"),
-            ]),
+            ],
+            path: "Sources"
+        ),
         .testTarget(
             name: "TurboNavigatorTests",
-            dependencies: ["TurboNavigator"]),
-    ])
+            dependencies: ["TurboNavigator"],
+            path: "Tests"
+        )
+    ]
+)
